Refactor routing implementation in graph package: introduce edge structure, update node methods to use edges instead of routes, and add configNode functionality.
