+++ internal/graph/debug_node.go
@@ -4,9 +4,10 @@ import (
 	"errors"
 	"fmt"
 
-	f "github.com/morphy76/lang-actor/pkg/framework"
 	g "github.com/morphy76/lang-actor/pkg/graph"
 )
+
+var staticDebugAssertion g.DebugNode = (*debugNode)(nil)
 
 type debugNode struct {
 	node
@@ -17,33 +18,14 @@ func (r *debugNode) OneWayRoute(name string, destination g.Node) error {
 	r.lock.Lock()
 	defer r.lock.Unlock()
 
-	if len(r.routes) > 0 {
+	if len(r.edges) > 0 {
 		return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("debugNode node [%s] already has a route", r.Name()))
 	}
 
-	r.routes[name] = route{
+	r.edges[name] = edge{
 		Name:        name,
 		Destination: destination,
 	}
 
 	return nil
-}
-
-type configNode struct {
-	node
-}
-
-// OneWayRoute adds a new possible outgoing route from the node
-func (r *configNode) OneWayRoute(name string, destination g.Node) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route [%s] from the config node [%s]", name, r.Name()))
-}
-
-// TwoWayRoute adds a new possible outgoing route from the node
-func (r *configNode) TwoWayRoute(name string, destination g.Node) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route [%s] from the config node [%s]", name, r.Name()))
-}
-
-// ProceedOnFirstRoute proceeds with the first route available
-func (r *configNode) ProceedOnFirstRoute(mex f.Message) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route from the config node [%s]", r.Name()))
 }
