+++ internal/graph/edge.go
@@ -4,7 +4,7 @@ import (
 	g "github.com/morphy76/lang-actor/pkg/graph"
 )
 
-type route struct {
+type edge struct {
 	// Name of the route
 	Name string
 	// Destination of the route
