+++ internal/graph/node_builders.go
@@ -14,10 +14,10 @@ import (
 
 func newNode[T any](task f.Actor[T], address url.URL) *node {
 	return &node{
-		lock:   &sync.Mutex{},
-		routes: make(map[string]route, 0),
-		actor:  task,
-		name:   fmt.Sprintf("/%s%s", address.Host, address.Path),
+		lock:  &sync.Mutex{},
+		edges: make(map[string]edge, 0),
+		actor: task,
+		name:  fmt.Sprintf("/%s%s", address.Host, address.Path),
 	}
 }
